requirements.txt
numpy==1.26.4
tensorflow==2.16.1
streamlit==1.36.0
Pillow==10.3.0
matplotlib==3.9.0
seaborn==0.13.2
